Exception info fix
git-svn-id: http://svn.vig/supra7/trunk@2089 f05b830f-2085-0410-a479-8f75d39d0b34

+++ src/webroot/cms/content-manager/page/PageAction.php
@@ -497,9 +497,11 @@ class PageAction extends PageManagerAction
 			try {
 				$page->moveAsLastChildOf($parent);
 			} catch (\Exception $e) {
-				$this->entityManager->remove($page);
-				$this->entityManager->remove($pageData);
-				$this->entityManager->flush();
+				if ($this->entityManager->isOpen()) {
+					$this->entityManager->remove($page);
+					$this->entityManager->remove($pageData);
+					$this->entityManager->flush();
+				}
 				
 				throw $e;
 			}
